+++ Octokit.Reactive/IObservableAuthorizationsClient.cs
@@ -12,9 +12,9 @@ namespace Octokit.Reactive
         IObservable<IReadOnlyCollection<Authorization>> GetAll();
         [SuppressMessage("Microsoft.Naming", "CA1716:IdentifiersShouldNotMatchKeywords", MessageId = "Get",
             Justification = "It's fiiiine. It's fine. Trust us.")]
-        IObservable<Authorization> Get(long id);
-        IObservable<Authorization> Update(long id, AuthorizationUpdate authorization);
+        IObservable<Authorization> Get(int id);
+        IObservable<Authorization> Update(int id, AuthorizationUpdate authorization);
         IObservable<Authorization> Create(AuthorizationUpdate authorization);
-        IObservable<Unit> Delete(long id);
+        IObservable<Unit> Delete(int id);
     }
 }
